(platform) Add new interfaces for location update
This change is to add new interfaces for location update.
An application can request a location update with a wait-time at any
time and receive the result through a listener interface. It also
provides a method to cancel a previously requested location update.
This implemenation is responsible for updating the current UE's location
using the fastest and accurate location provider.

Bug: 381525496
Bug: 388625320
Flag: NONE pre-production
Test: atest ImsStackJavaTests, ImsStackNativeTests
Test: manual (verified through test code that can receive location
update request and its result)

Change-Id: I9c97cd88c4cb5bc2c7d21f2482dbb52680fb43b1