(sipcore) Retry INVITE transmission on TCP connection closed
This change is to add an exception handling for the scenario where the
TCP connection is closed by the network while the SIP INVITE request is
being sent.
If an INVITE request is sent via TCP transport and the TCP connection is
closed before a response is received for a certain time, the INVITE
request is retransmitted over a new TCP connection.
This measure aims to increase the success rate of IMS call setup.

Bug: 361223640
Flag: NONE pre-production
Test: manual (verified INVITE retransmission for the scenario where the
TCP connection is closed by the peer while the SIP INVITE request is
being sent)

Change-Id: I378572e4686d2d0770b070c91800f13e3dc610fc