(Mtc) Skip user-connected operation if subscription failed
This fix addresses an issue where the MTC incorrectly waits for the
conference event package (CEP) result even when the conference event
subscription fails.

Specifically, when the 'conference_sip_flow_order_int' value is
2(CONFERENCE_SIP_FLOW_SUBSCRIBE_AND_NOTIFY_REFER), the MTC sends a
REFER and then waits for the CEP result before processing to the next
operation. This was causing a problem because if the subscription
failed, the MTC would still wait indefinitely for the CEP result and
would not proceed with the subsequent actions.

With this fix, if the subscription fails, the MTC will now only
consider the result of the NOTIFY for REFER (R-NOTIFY) and proceed to
the next operation accordingly, without waiting for the CEP result.

Test: atest ImsStackNativeEnablerMtcTests
Test: initial merge operation with sub failure
Test: adding a participant with sub failure
Bug: 413255273
Flag: NONE pre-production
Change-Id: I3d461db632ffb890aec5c67f24c3f03980614349