(core) Adjust Session event sequence in abnormal case (re-INVITE/ACK)
This change makes the Engine behave as if an ACK request has been
received when a re-INVITE request is received during the initial session
setup phase, before the actual ACK request arrives. At this point, the
Engine proceeds with calling the SessionUpdateReceived listener.
However, to ensure the correct event sequence, the Engine adjusts the
sequence so that the SessionUpdateReceived listener is called after the
SessionStarted listener.

Bug: 420533123
Flag: EXEMPT bugfix
Test: manual (verified normal handling when a re-INVITE request is
received before the ACK request arrives during the initial session
setup)

Change-Id: Ib338ff764869dfb610a83421b49e6f4693bd3cae