(sipcore) Recreate SIP socket if it is closed or being closed
This change is to add the status check of SIP socket before sending a
SIP message. If the socket is already closed or in the process of
closing, a new SIP socket will be created to enable the transmission of
the SIP message.

Bug: 417847668
Flag: EXEMPT bugfix
Test: manual (verified the behavior during handover scenarios where the
socket is closed timing-wise, using the carrier test case)

Change-Id: I2f577909ec1bc7a4b008d0969a4b391df2ecbf88