(Mtc) Fix: incorrect CallType retrieval
This commit fixes an issue where an incorrect origianl CallType was
being retrieved in the following scenario:
- Scenario: Receiving a 200 OK with port number 0 in response to an
  upgrade request.
- Issue: MTC incorrectly judged that the CallType had changed, leading
  to erroneous behavior, even though it had not.

Currently, `UpdatingInfo#IsRequestedModifying()` returns `false` because
`IMtcSession#GetPreviousCallType()` retrieves the target CallType
instead of the original CallType when a 200 OK with port 0 is received.

To address this, the code has been modified to always store the origianl
CallType by retrieving the value held by `MtcSession` before
`UpdatingInfo` is created.

Test: outging / incoming upgrade test
Test: atest ImsStackNativeEnablerMtcTests
Bug: 419322850
Flag: NONE pre-production
Change-Id: I7904161775d183edc46a8db119ed4b5da737a681