(Aos) Changed to notify VopsChanged if only Plmn is changed
Currently VoPS change is not notified if the new Vops state is same.

However, it should notify when only Plmn is changed as well to ensure
integrity of the VoPS and Plmn pair.

This CL changes DcNetWatcher to call the VopsChanged listener callback
if only plmn is changed as well.

Bug: 418080869
Test: gTest and Manual
Flag: NONE pre-production
Change-Id: Iaabee0b3d83146533fbec2b30558ef2ff111f52f