(Ssc) Update NoReplyTimer if exists in XML
This CL handles the abnormal case where NoReplyTimer did not exist when
the UE first queried the XCAP server, but the server was updated with a
NoReplyTimer later. As a result, the UE was unaware of the existing
node and subsequently failed to update NoReplyTimer because it tried to
insert a new node instead of modifying the existing one. With this fix,
UE will update the NoReplyTimer instead of trying to insert.

Bug: 424007251
Test: atest SscXmlCreatorTest
Test: manual test on TMUS network
Flag: NONE pre-production
Change-Id: Ice4a48480530e0a6233b0c935a7bd82034559536