(SIP) Update RPR transaction timeout handling
This change is to enable the SIP protocol stack to differentiate and
process RPR transaction and INVITE transaction separately when a SIP
transaction timeout occurs. This modifies the Enabler to send a failure
response for the corresponding INVITE request when a PRACK is not
received after an RPR is sent.

Bug: 232669991
Flag: EXEMPT bugfix
Test: manual (verified RPR transaction timeout case with TISS)
Change-Id: I50f7a04c50b1d1fa15b41ab0a94dc97aeb8afdca