Update location/address once when IMS service is available
This change is to obtain the location information for PIDF-LO once
ImsStack is running and service is available and use the cached address.
This allows PIDF-LO information to be obtained and used normally in the
carrier cert-tests using the closed network and included in the SIP
message body part.

Bug: 418126301
Bug: 417329879
Flag: NONE pre-production
Test: atest ImsStackJavaTests
Test: manual (verified that the location is obtained at the start stage
of the service and carrier's cert-tests)

Change-Id: I14106c891a8f98c236658c9f5b935617fee256d3