(Mtc) Fix to use stale UPDATE method in case of renegotiating
This change ensures that when ISession's GetState() returns STATE_RENEGOTIATING, the GetPreviousResponse() method is called using SESSION_STATE_UPDATE instead of SESSION_UPDATE.

When a re-INVITE is sent and a success/failure response is received, leading to ISessionListener#SessionUpdated() or ISessionListener#SessionUpdateFailed() being called, the response message is retrieved and processed. At this point, a race condition can occur:
- if a new incoming re-INVITE has already been received, the engine manages the last SESSION_UPDATE information based on that received re-INVITE.
- Consequently, if the response is fetched using the SESSION_UPDATE method in this scenario, it will return null.
Since the session's state becomes STATE_RENEGOTIATING in this case, the enabler can use this condition to obtain information about the previously sent re-INVITE and its respnoses.

Test: atest ImsStackNativeEnablerMtcTests
Test: simulated normal and stale update scenarios
Bug: 420593302
Flag: NONE pre-production
Change-Id: I896363f0b0294a18dd8780ffa90c90a977a024a5