Fix to invoke SuppServiceReceived by onCallUpdated
This adds notifyCallSessionMultipartyStateChanged() into onCallUpdated.
notifyCallSessionMultipartyStateChanged() is required by upper layers to
change the UI to conference call when a 1-to-1 call is extended to a
conference call.
Currently, onCallUpdated() omits this logic so the call cannot be
transitted to a conference call when re-INVITE with 'downgrade' is
received from the conference focus.

Test: atest ImsStackJavaTests:ImsCallSessionImpleTest
Test: verified the issue scenario
Bug: 424245475
Flag: NONE pre-production
Change-Id: Icf241d38d0095bd15ca0794fb24d0cc3f2434212