Use cached address with validity period for PIDF-LO
Currently, LocationAgent caches Location information and verifies and
uses Location information based on the validity period.
Similarly, this change adds the ability to store Address that is
resolved based on Location and use it based on the validity period in
order to increase the probability of acquiring PIDF-LO content.

Bug: 382588547
Flag: NONE pre-production
Test: atest ImsStackJavaTests
Test: manual (verified the cached Location and Address are properly
loaded even if the device reboots with APM ON)

Change-Id: I4e773057cf6ae5bf62257c1a87de746aacbea709