(MTC)(DTAG) terminates the call upon receiving 503
DTAG MTR-40342
Upon receiving a 503 response with a Retry-After header to an initial INVITE request,the UE must not automatically resend  the INVITE when the Retry-After timer expires. Any new call attempt should be treated as a completely new session, requiring user action to initiate.

to support MTR-40342, There is no more redial operation for INVITE(No#1) when receiving 503, this call INVITE(No#1) is simply processed as a call fail,and the Retry-After timer will not run, since the call that is restricted during the Retry-After time is only INVITE(No#1) (which is already terminated), and the remaining newly in progress calls (ex, INVITE(No#2) or INVITE(No#3) are interpreted as being allowed.

Test: TISS simulation (INVITE - 503 w/ Retry-After)
Test: atest ImsStackJavaTests
Test: atest ImsStackNativeTests
Bug: 427619676
Flag: NONE pre-production
Change-Id: I61b8de0101bddd7e62444affb7815d814c83504e