Get the data network type regardless of its registration state
When making an emergency call in a VoWiFi registered state, the network
type was not correctly identified, leading to a phenomenon where audio
transmission/reception failed even if the call was connected.
This change is to directly read the actually stored network type
regardless of its registration state when reading the PS network
registration information's network type. This ensures that the network
type can be correctly identified even if the network is in a limited
service state.

Bug: 427309511
Flag: EXEMPT bugfix
Test: atest ImsStackJavaTests
Test: manual (verified the PS(cellular) emergency call while VoWiFi is
registered)

Change-Id: Icdfa5d6bb6a908de76141044278e50cd120f410c