(Aos) cppcheck: Avoid virtual call in AosSubscriberManager constructor
The virtual functions IsIsim() and IsUsim() were being called from
ConfigureAsDefault(), which is executed during the construction of
AosSubscriberManager.
This triggers a [virtualCallInConstructor] warning because dynamic binding is not guaranteed to work correctly before an object is fully constructed. This change resolves the warning by directly accessing the member variables m_bIsim and m_bUsim instead, which is safe and reflects the actual behavior during construction.

Bug: 388646581
Flag: NONE pre-production
Test: atest ImsStackNativeTests

Change-Id: I243cea76d5303ab74446af25eeca2b9e9cdb0749