(core) Add a preview mode for SDP offer/answer exchange
This change is to enable early media transmission by performing SDP
negotiation when a non-RPR message containing SDP is received.
By default, if the SDP offer/answer state is in preview mode, SDP
negotiation is not performed for subsequent non-RPR, RPR, and successful
final response messages even if SDP exists.
However, depending on the carrier-config, if a non-RPR or RPR message
containing SDP is received in the preview mode, a new SDP negotiation
may be performed as an SDP offerer.
Also, if a new SDP offer is sent or received in SDP preview mode, the
preview mode will be stopped.

Bug: 420328481
Flag: NONE pre-production
Test: atest ImsStackNativeTests
Test: manual (verified SDP preview state scenario with TISS)
Change-Id: I38d6e54c9adc2de81073b9e78ae57a35de9b46bf