Save and use a config of last inserted SIM
This change introduces a new policy for determing which carrier
configuration to use when booting a device without SIM card.
A new optional config item is defined, allowing the ImsStack to operate
based on the configuration of the last inserted SIM. This involves
saving the most recent configuration and enabling its use upon reboot,
even if no SIM is present.
In SIM LOCKED case, the cached config can only be used if the SIM serial
ID is the same.

Bug: 424310036
Bug: 430705663
Flag: EXEMPT bugfix
Test: atest ImsStackJavaTests
Test: manual (verified the carrier configuration based on the SIM state
when the device reboots)

Change-Id: I6a16eeeb273e020db631184deb407df7bdc57820