(Aos) Update available P-CSCF address even if IP version is changes
When the P-CSCF address changes, only the P-CSCF address with the same IP version as the current P-CSCF's IP version was considered as a possible address.
Therefore, If the IP version of P-CSCF is changed, the new P-CSCF address is not updated.

This CL has been modified to update to a new P-CSCF even if the IP versions are different if there is no P-CSCF with the same IP version as the current P-CSCF. (However, it is updated only if there is a Link/P-CSCF address corresponding to the changed IP version.)

Bug: 430449560
Test: atest ImsStackNativeTests and DcApnTest
Test: Equipment testing that reproduces the problem situation
Flag: NONE pre-production
Change-Id: Iec92bd88a816d55d1a180420d554d47ec9d08336