(Mtc) Temporarily override network country ISO for testGetNetworkCountryIso
The testGetNetworkCountryIso test is failing because it's unexpectedly
retrieving country ISO from ImsPrivateProperties.Persistent. This commit
implements a temporary fix to provide a valid country ISO at the start
of the test, allowing it to pass. A more robust solution will involve
mocking SharedPreferences to ensure complete test isolation. This
temporary fix will be removed once that improvement is in place.

Test: atest ImsStackJavaTests
Bug: 430729792
Flag: NONE pre-production
Change-Id: I358f561900a4240ffd07f0c6f3a6e02b10b7db79