(AOS) Check holding VOPS state when deciding if process VOPS changes is needed
Problem: The issue occurs during an active call: when the VOPS status changes to false, the state is held to be processed later. However, if VOPS recovers to true before the call ends, this holding VOPS state is not correctly updated. As a result, the holding VOPS state false is acted upon after the call, causing BLOCK_VOPS to be added to the mobile block list. In addition, this makes the gims in a stuck state where future VOPS state updates are ignored, making this BLOCK_VOPS persistent.

bug: 430771561

Test: atest ImsStackNativeTests/ImsStackJavaTests
Test: LTP test PASS
Flag: NONE pre-production

Change-Id: Idac1cd12fafb88d1f2bb46fef33ea208b539ae59