Add private config commands for internal/automation test
This change provides a way to read and write the internal carrier
configuration to support automated testing of internal test or
carrier's certificate test.
These commands will be officially available for internal and field test
to update the internal carrier configuration instead of hidden menu.

These commands are based on the broadcast intents and are defined as
follows:
 - SET_CONFIG: update the configs (overriding existing configs)
 - CLEAR_CONFIG: restore the configs to original configs
 - GET_CONFIG: read the current configs
 - NOTE: it's only available in non-user build.

Bug: 430014581
Flag: NONE pre-production
Test: atest ImsStackJavaTests
Test: manual (verified set/clear/get commmands on target device based on
the logs and config dump)

Change-Id: I91da43e82be7c7789acfb7355882b0f196d1cca5