(config) Handle invalid EF_domain by using IMPU
An invalid EF_domain ("0") from the ISIM results in an invalid phone-context in the Request-URI. Consequently, the server does not respond, leading to a VoLTE call failure.
This change introduces a fallback mechanism to handle this specific case.

Changes:
The logic to handle an invalid EF_domain ("0") is now implemented directly within the SubscriberConfig::UpdateHomeDomainName function.

Bug: 369487367
Test: atest ImsStackNativeTest
Flag: NONE pre-production
Change-Id: Ic3a4be5d1d7c76a08db54b9f0e2fcc838cb33484