(Aos) Shouldn't block PLMN when IMS service disabled
PixelCare module could toggle IMS service to recover the device state
when it has some problem with IMS registration. In the some cases, the
IMS service can be disabled even when IMS is registered. If IMS service
is disabled, Aos performs IMS de-registration and checks if the current
PLMN should be blocked or not. In case of IMS service disable, the PLMN
shouldn't be blocked because it's not a problem with the network or IMS
core. This CL is to not block the PLMN for normal IMS de-registration
cases.

Bug: 432148813
Test: atest AosApplicationTest
Test: Manual test on TMUS live network - disable IMS service
Flag: NONE pre-production
Change-Id: Idc9c8f266847b7a57ba4a63a64170cd9a4a392f5