(core) Add a new method for aborting early UPDATE transaction
If 200 OK response for an INVITE is received while an early UPDATE
transaction is still pending, the session is considered fully
established. The precise handling of this UPDATE transaction during
early media SDP negotiation can vary based on the carrier's
requirements.
Therefore, this change introduces a new method that allows the Enabler
to abort the early UPDATE transaction as per the carrier's requirements.

Bug: 411385901
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verified early UPDATE transaction is successfully aborted
and the subsequent UPDATE transactions are worked as expected)

Change-Id: Idd49849b8ac70bf2b84c73aaa02e37419013735a