(Aos) Ignore barring info from NR cells using getCellIdentity
This change utilizes the new BarringInfo#getCellIdentity() API to prevent the IMS stack from incorrectly applying barring information received from a neighboring NR cell when the device is on a different RAT (e.g., LTE).
The handleBarringInfo method is updated to check the type of the CellIdentity. If it is an instance of CellIdentityNr, the barring information is now ignored, and the SSAC info is not updated.
This ensures that SSAC is only applied based on barring information relevant to the device's active network, preventing erroneous PLMN blocking and improving service stability.

Bug: 431881770
Test: ImsStackNativeTest and manual test
Flag: NONE pre-production

Change-Id: Icea79f4ffa6de371dc66f5539dfd4a7ca4a985d8