Fix to set the proper negotiated video SDP fmtp attributes.
[Issue]
The video SDP negotiation is to follow the peer side resolution in the HEVC codec case.
The video SDP negotiation is to follow the local side level even the peer answers the lower level.

[Fix]
Modify to negotiate the video SDP to follow the local side resolution in the HEVC codec.
Modify to negotiate the video SDP to follow the peer side level when it is lower than local's.

Bug: 433580569
Bug: 424245475
Bug: 435260312
Test: atest ImsStackNativeEnablerMediaTests
Flag: EXEMPT bugfix
Change-Id: I46bdf45e12ad1e26032a83c9e0dd7d63edda2c97