(Mtc) Play local tone when 180 response is received without negotiation
The device is configured to use NW tone but SDP is not negotiated
when 180 response is received. In this situation, the device does
not start the waiting timer for NW tone and could not use local
tone as well.

This commit update code to forcibly play local tone for the
situation.

Test: atest ImsStackNativeTests
Test: Verify code change in the live network
Bug: 433355433
Flag: NONE pre-production

Change-Id: Ib4a2dfcefabd7a4ea315f79b9d3b983f0d25137f