Modify to update preferred IP address when IP changed.
For IP addresses of the emergency type, the preferred IP address was updated only when the cached address was changed.
However, if an IPv6 address was allocated very quickly after an IPv4 address was allocated, causing the IPv6 address to be cached after the IPv4 address was selected as the preferred IP address.
To handle these exceptional scenario, we've modified the behavior so that when an IP changed event occurs, the preferred IP address is updated even if the cached address hasn't changed.

Test: atest ImsStackNativeTests
Bug: 431143307
Flag: NONE pre-production
Change-Id: Ib6c5711474b5f9f202f3501c8517725af7a8ddff