Fix the payload number mismatch
[Issue]
When the video SDP negotiated from the multiple payloads, and the local payload type number is mismatched with the negotiated one, the incorrect payload number is set to the video config which causes the no video frames to the other party.

[Fix]
Fix to set the negotiated payload number to the video config.

Bug: 435260312
Test: atest ImsStackNativeEnalerMediaTests. Verified by the video call in the live network including the video conference call.
Flag: EXEMPT bugfix
Change-Id: I9626848e23571b56a4de2a4a49240b2554212efd