(AoS) clear SIP RT config when sending initial registration
The fix addresses an issue where an incorrect SIP header was being
set because existing SIP RT config setting is not cleared when the SIM was
changed. To resolve this, the modification was made to clear the SIP RT
config setting at the time of the initial registration.

Flag: NONE pre-production
Bug: 434779163
Bug: 440243369
Test: manual and atest ImsStackNativeEnablerAosTests
Change-Id: I8003da07cdc850394f351359599db6befda552c2