(core) Fix incorrect preview mode behavior for offerless INVITE
When receiving an offerless INVITE request and performing an SDP
offer/answer exchange with a 183 response and PRACK request, the Engine
would interpret this as preview mode and it causes incorrect actions in
the Enabler layer.
This change prevents incorrect preview mode detection for MT session by
determining whether to recognize preview mode based on the
directionality of the session start.

Bug: 437765274
Flag: NONE pre-production
Test: manual (verify the handling of offerless INVITE request -
VoIP 3.20.4)
Test: manual (verify the preview mode test cases with TISS)

Change-Id: I596379224f79e00010a17d13d65862fd5125db65