(Mtc) Auto missed call is not processed correctly
The Google IMS notified the Telephony of an auto missed call and waited for a specific API to be called from the Telephony for a certain period of time.
However, in this case, the specific API was not called within that period. As a result, the auto missed call was not disconnected gracefully.
To resolve this, we need to ensure that the specific API is called.
For more details, please refer to the document in the Bug.
One-pager : http://docs/document/d/1tIr3nRg0xbjlS1qu1MNxSuYbqG4L0n--XzfJP-8zpys?pli=1&tab=t.0#heading=h.xgjl2srtytjt

Test: atest ImsStackJavaTests
Test: simulated with TISS
Bug: 438800185
Flag: NONE pre-production
Change-Id: I33d3955127c15add966695fbc895c0f27741522c