Add ImsStackNativeTests to device-tests suite
This change modifies the Android.bp file for the ImsStackNativeTests
module to include "device-tests" in the test_suites array within the
imsstack_native_test_defaults block.

Reasoning:
The test_suites property in Android.bp dictates which test collections a
module belongs to. For a test to be packaged into the device-tests.zip
artifact, which is used for running tests on actual Android devices via
frameworks like Trade Federation and ATP (Android Test Production), it
must be included in the "device-tests" suite.

Previously, ImsStackNativeTests was only in "general-tests", which does
not guarantee its inclusion in the device-tests.zip artifact. This
prevented ATP from finding and executing these native tests on a device,
leading to "Could not find test file" errors.

By adding "device-tests", we ensure that the ImsStackNativeTests
binaries are correctly placed within the target/ directory of the
device-tests.zip artifact during the build process, making them
available for on-device testing.

Package build script rules :
https://source.android.com/docs/core/tests/development/test-mapping?hl=ko#package-rules

Bug: 424373997
Flag: TEST_ONLY
Test:ATP test
Change-Id: I5c9ca94af7a060e79d085dc6216495c19cc9e5bf