(sipcore) Fix a crash while accessing SIP socket
During the emergency call setup, the normal IMS registration performed
the de-registration because "vops" was not supported. When clearing the
SIP socket due to IMS de-registration, an ACK request was sent by
referencing a SIP socket object that had already been freed, which
resulted in a crash.

To fix this issue, the following changes are made:
- When cleaning up SIP sockets, it checks for the specified local IP
  address and only clears the sockets that match it.
- When a SIP socket object is validated in SIP transport layer, it
  confirms the actual existence of the corresponding object.

Bug: 441751752
Flag: NONE pre-production
Test: manual (verify SIP socket destruction after IMS is de-registered)
Test: manual (verify IMS registration/call on target device)
Change-Id: Id23772a625e91ae814a3f7f072542fc1fe970889