(mtc) Terminate call on unsupported network change in MTC
This change introduces logic in the MTC enabler to proactively terminate
an ongoing call when the device moves to a network that does not support
service continuity.

It makes the device immediately terminates an ongoing 5G call when it's
moved to the 2G/3G network that doesn't support transition, avoiding a
long delay before the call drop due to media timeouts.

Test: atest ImsStackNativeEnablerMtcTests
Test: TISS and H/O test command (INTENT_MTC_TEST)
Bug: 441592101
Flag: NONE pre-production
Change-Id: I67303751d180a0751f30c5f9c48ff065a8bd4e26