(Mtc) `StartImsTraffic` and `StopImsTraffic` pair are not working correctly
The cases below should be satisfied, but are not satisfied now. So, correct this.
Case. StartImsTraffic > OnConnectionSetupPrepared > call1 connected > add same traffic call2, call2 connected, call1 end > call2 end > StopImsTraffic
Currently, with this case, the traffic is released when call 1 is released even if the call 2 still uses it because subsequent calls aren't added to the traffic's list if the traffic is already activated.

Test: atest ImsStackNativeEnablerMtcTests
Test: simulated with TISS
Bug: 433869994
Flag: NONE pre-production

Change-Id: I0dc3d26dbca7aba4ea8a9ddd8506941d87d382d1