(mtc) Set update type when receiving an incoming update
Currently MTC doesn't update the UpdateType of UpdatingInfo when
receiving an incoming update. So ProcessingBlockRule can't properly
block incoming calls when the incoming update is progressing (e.g.
call type changing).

This CL makes UpdateType to be set properly in this case.
If both call type and hold resume are changed, it's set to SESSION.

Test: TISS simulated the bug scenario
Test: atest ImsStackNativeEnablerMtcTests
Bug: 444319068
Flag: NONE pre-production
Change-Id: I8ff4c0347e7ba1df7406dd3e21aa476f2cd909a0