(Aos) Notify IAosSubscriber::NOT_READY when UICC is removed
If UICC is removed, IMS/EIMS registration should be blocked by
BLOCK_SUBSCRIBER_INCOMPLETED reason. If the UICC previosuly inserted is
carriers that use ISIM for IMS registration, it's handled properly.
However, if the UICC was for carriers that use USIM for IMS
registration, it's not blocked. This CL is to fix it by handling SIM
state change events especially for ABSENT state.

Bug: 444393978
Test: atest ImsStackNativeEnablerAosTests
Test: manual test - SIM inserting and removing
Flag: NONE pre-production
Change-Id: I76f19a861a73049c2e201cf90664d4e0174c9fca