(Aos) Release EPDN when anonymous ecall ended on the specific networks
EPDN should be controlled based on the carrier's requirements when the
emergency call ends. However, there are some specific conditions where
UE cannot load the carrier configurations, such as UICC is not present
or PIN-locked UICC is equipped. This CL is to control EPDN based on PLMN
of the network that UE made the emergency call on, specifically under
the no-UICC and PIN-locked UICC conditions.

Test: atest AosEApplicationTest
Flag: NONE pre-production
Bug: 445098633
Change-Id: I99b25a5e5eae50e54ffcc9bfa35b7c9ef5578da9