(Aos) Add null check for Handler in AosService.ListenerProxy
The `AosService.ListenerProxy.onMessage()` method calls `mHandler.sendMessage()` to deliver messages from native code back to the service thread. If the `mHandler` object is null when this method is called, it results in a `NullPointerException`.

This commit fixes the potential crash intermittently when enabler re-start by Adding an explicit null check for `mHandler` at the beginning of `ListenerProxy.onMessage()` and returning early.

Flag: NONE pre-production
Bug: 447005675
Test: atest AosServiceTest
Change-Id: Idb9b3b42b18a5f944fd4c2671f20e5c578d66d93