(MTS) Notify e-SMS state to AOS
This CL introduces a new notification system that provides AOS with
direct and immediate information about the status of an emergency SMS
(e-SMS) transaction from the MTS.

The new notification, `NotifyEmergencySmsState`, is required to address
critical "race conditions" where an emergency SMS is sent or received at
the same time the phone's emergency registration is about to expire.
This allows AOS to proactively refresh the registration with the
network, ensuring the message can be sent successfully.

Bug: 289441507
Doc: go/mts-notify-esms-aos
Test: Manual MO/MT SMS
Test: atest ImsStackNativeTests
Flag: NONE pre-production
Change-Id: I343e3d4b5a415519f6d614fe2f787afe6300bd53