(Aos) Refactor mock expectation setup in AosNConfigurationTest
This change refactors the mock expectation setup from `EXPECT_CALL` with `WillOnce` or `WillRepeatedly` to using `ON_CALL` with `WillByDefault`.

Using `ON_CALL` is appropriate for defining the stubbing behavior of a
mock object before the function call under test, which improves the
clarity of the test intention. It explicitly sets the default behavior, while `EXPECT_CALL` is reserved for verifying that a function was called.

Flag: NONE pre-production
Bug: 427733661
Test: atest ImsStackNativeEnablerAosTests
Change-Id: Ib063778183fec103249111fd4b6a4b38ce33dd02