(mtc) Use default reason phrase for 488 when it's not about codec support
Verizon requires special reason phrase for 488, but it's restricted to
the following condition.

VZ_REQ_5GNRSAVOICEVIDEO_4105999311948844 If the device does not support
any of the codecs listed in the SDP offer, ... send SIP 488 with reason
phrase "Codec Not Supported".

This CL corrects maps from NegotiationResult to CallReasonInfo's code
and extra, making the reason phrase and warning header are set as
expected.

Test: TISS simulation
Test: atest ImsStackNativeEnablerMtcTests
Bug: 447238172
Flag: NONE pre-production
Change-Id: Ia2080ae0bbd3e8389f115696dc1ec196e6579f7c