(Aos) Resolve varargs call ambiguity in TestRegistration
A "non-varargs call of varargs method" warning was triggered when passing a Uri[] array to the EventRecord constructor.

This warning occurs due to an ambiguity: the Java compiler is unsure whether to treat the array as a single argument or to expand its elements into multiple arguments for the varargs parameter (Object...).

By explicitly casting the array to (Object), we instruct the compiler to treat the entire array as a single object. This resolves the warning and correctly reflects the intent to store the array itself within the event record.

Bug: 424347544
Flag: EXEMPT TEST_ONLY
Test: atest
ImsStackTest: com.android.imsstack.its.tests.registration.tests.PrimaryRegistrationTest

Change-Id: Idee66fdffe9175c56c72cacc6178debbbe0c3339