(Mtc) Fix unexpected SDP in PRACK
Issue: PRACK for a forked 18x contains unnecessary SDP
In case of the precondition feature is not supported by the remote UE,
no QoS precondition confirmation is required. But, currently the remote
precondition supportability is updated after sending PRACK in some
condition so the QoS precondition confirmation is added to the PRACK
message.
This change modifies UE to update the remote capability before sending
PRACK so that the PRACK message correctly reflects the remote
capability.

Bug: 447519886
Test: atest ImsStackNativeEnablerMtcTests
Test: simulation by TISS
Flag: NONE pre-production
Change-Id: I2016d365e687abaa86dc73b2821d395a17728512